+++ src/features/resource-calendar/components/ilamy-resource-calendar/ilamy-resource-calendar.tsx
@@ -4,10 +4,10 @@ import { ResourceCalendarBody } from './resource-calendar-body'
 import type {
   IlamyResourceCalendarProps,
   IlamyResourceCalendarPropEvent,
-  ResourceCalendarEvent,
 } from '@/features/resource-calendar/types'
 import { DAY_MAX_EVENTS_DEFAULT, WEEK_DAYS_NUMBER_MAP } from '@/lib/constants'
 import { safeDate, normalizeEvents } from '@/lib/utils'
+import type { CalendarEvent } from '@/components/types'
 
 export const IlamyResourceCalendar: React.FC<IlamyResourceCalendarProps> = ({
   events = [],
@@ -41,10 +41,9 @@ export const IlamyResourceCalendar: React.FC<IlamyResourceCalendarProps> = ({
 }) => {
   return (
     <ResourceCalendarProvider
-      events={normalizeEvents<
-        IlamyResourceCalendarPropEvent,
-        ResourceCalendarEvent
-      >(events)}
+      events={normalizeEvents<IlamyResourceCalendarPropEvent, CalendarEvent>(
+        events
+      )}
       resources={resources}
       firstDayOfWeek={WEEK_DAYS_NUMBER_MAP[firstDayOfWeek]}
       initialView={initialView}