+++ src/lib/utils/index.ts
@@ -45,7 +45,7 @@ export const omitKeys = <T extends object, K extends keyof T>(
  * const events = normalizeEvents<IlamyCalendarPropEvent, CalendarEvent>(propEvents)
  *
  * // Resource calendar events
- * const resourceEvents = normalizeEvents<IlamyResourceCalendarPropEvent, ResourceCalendarEvent>(propEvents)
+ * const resourceEvents = normalizeEvents<IlamyResourceCalendarPropEvent, CalendarEvent>(propEvents)
  */
 export function normalizeEvents<
   TInput extends {